Emergentica/requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-aws>=0.1.0
langgraph>=0.0.40
langsmith>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# AWS Bedrock
boto3>=1.34.0
botocore>=1.34.0

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.30.0

# Telephony & Voice
twilio>=8.11.0
httpx>=0.26.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
tenacity>=8.2.0
rich>=13.7.0

# Optional: Enhanced Search
# valyu>=0.1.0
